fix/review: menuOwner

+++ lib/models/ingredient.dart
@@ -1,9 +1,9 @@
 class Ingredient {
   String name;
   String units;
-  int amount;
+  double amount;
 
-  Ingredient(String name, String units, int amount) {
+  Ingredient(String name, String units, double amount) {
     this.name = name;
     this.units = units;
     this.amount = amount;
